fix(2oath):Refactor 2FA and JWT handling
The code changes include a refactor of the two-factor authentication (2FA) and JSON Web Token (JWT) handling. The condition for checking if 2FA is enabled has been made more explicit. Instead of checking for not true, it now checks for false explicitly.

In addition, the property name 'access_token' in the auth service has been changed to camel case ('accessToken') to maintain consistency in naming conventions across the application.

Lastly, removed unnecessary check in jwt-2fa strategy that was throwing an exception if payload indicated user was already authenticated with 2FA. This simplifies the logic and avoids potential confusion or errors due to redundant checks.