fix(friendship): Updated the method for counting followers and followings
The previous implementation only counted accepted and pending friendships as followers, but now it also includes those who have sent a friend request to the user. Similarly, for followings count, it now includes users that the current user has received a friend request from in addition to those they are following.